random_number_guess/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
joblib==1.4.2
numpy==2.2.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
urllib3==2.3.0
